ui: add nearest channel GPS editing

+++ openrtx/src/rtx/OpMode_M17.cpp
@@ -349,8 +349,10 @@ void OpMode_M17::txState(rtxStatus_t *const status)
         streamType_t type = {};
         type.fields.dataMode = M17_DATAMODE_STREAM;     // Stream
         type.fields.dataType = M17_DATATYPE_VOICE;      // Voice data
-        type.fields.encType  = encryptionActive ? encType : M17_ENCRYPTION_NONE;
-        type.fields.encSubType = encryptionActive ? encSubType : M17_META_TEXT;
+        type.fields.encType = encryptionActive ? encType
+                                               : static_cast<uint8_t>(M17_ENCRYPTION_NONE);
+        type.fields.encSubType = encryptionActive ? encSubType
+                                                  : static_cast<uint8_t>(M17_META_TEXT);
         type.fields.CAN      = status->txCan;           // Channel access number
 
         lsf.setType(type);
